pyproject.toml
[tool.poetry]
name = "pydantic-openapi-orm"
version = "0.1.0"
description = ""
authors = ["Anentropic <ego@anentropic.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pydantic = "^1.5.1"
PyYAML = "^5.3.1"
email_validator = "^1.1.0"  # for pydantic EmailStr fields
jsonref = "^0.2"
httpx = "^0.12.1"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
hypothesis = "^5.14.0"
hypothesis-jsonschema = "^0.12.1"
ipython = "^7.14.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
